third_party/blink/web_tests/fast/forms/datetimelocal-multiple-fields/datetimelocal-multiple-fields-change-layout-by-value-expected.txt

Change multiple fields datetime-local input UI layout by value

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS widthOfEmptyValue < widthOfOverflowValue is true
PASS widthOfEmptyValue is widthOfValidValue
PASS widthOfEmptyValue < widthWithSecond is true
PASS widthWithSecond < widthWithMillisecond is true
PASS successfullyParsed is true

TEST COMPLETE
